requirements.txt
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.75
botocore==1.35.75
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Django==5.1.3
django-crispy-forms==2.3
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
future==0.18.3
idna==3.10
jmespath==1.0.1
numpy==1.26.4
packaging==24.2
pillow==11.0.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
